Here's a suggested commit message:
Remove docs icon from Alfred workflow

Removes the unused docs.png icon file from the Alfred workflow directory.
This icon was likely part of a documentation-related workflow feature that
is no longer needed or has been replaced.

Since this is a binary file deletion and appears to be an icon resource,
the impact should be minimal unless other parts of the workflow explicitly
reference this image file.

Note: If any workflow configurations depend on this icon file, they will
need to be updated to prevent broken references.